Add logging infrastructure
Code in /src/main uses just the SLF4J API for when the
project built from this template is a library, which
should rely just on a logging API, not an implementation.

The code in /src/test though, has access to the Logback
implementation of the SLF4J API. So logging will work
when running the tests.

Non-library projects can, and should, change the logback
dependencies scopes in build.scala so that they can use
Logback in production code as well.